Incremental shooter speed
Reworked FuelShooter to use incremental speeds. Now saves the current speed and uses commands to increment or decrement the speed by 5%. Can also be set directly if needed. stopShooter does not set the speed to 0, it just pauses the shooter and it can be reactivated at the last used speed.